feature(act-runner): convert deployment into a stateful set (#9)
* feature(act-runner): convert deployment into a stateful set

This change replaces the Deployment with a StatefulSet to support the
use of volumeClaimTemplates. Each runner now gets its own persistent
volume, which is necessary for storing credentials and identity
information. Without this, runners share the same data and only one
shows up in the Gitea Actions UI due to the sharing of the same IDs and
tokens.

The option to provide an existing PersistentVolumeClaim has been
removed, as it no longer fits the per-runner volume model.

* fix(actions): don't run release workflow for PRs

Signed-off-by: Filipe Utzig <filipe@gringolito.com>